requirements.txt
Django==5.2.5
gunicorn==22.0.0
whitenoise==6.7.0
numpy==2.1.2
Pillow==10.4.0
psycopg[binary]==3.2.1
scikit-learn==1.5.2
torch==2.4.1
